[운영배포] 1.별->달자동교환 2.환전 별 타입 사이즈 변경

+++ src/main/java/com/dalbit/member/vo/procedure/P_ExchangeApplyVo.java
@@ -9,12 +9,13 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.servlet.http.HttpServletRequest;
+import java.math.BigDecimal;
 
 @Getter @Setter
 public class P_ExchangeApplyVo {
 
     private String mem_no;
-    private int byeol;
+    private BigDecimal byeol;
     private String account_name;
     private String bank_code;
     private String account_no;
